+++ Sources/SoundIODemo/sine/main.swift
@@ -18,7 +18,16 @@ func main() throws {
         let layout = outstream.layout
         let secondsPerFrame = 1.0 / Float(outstream.sampleRate)
 
-        try! outstream.write(frameCount: frameCountMax) { (areas, frameCount) in
+        var framesLeft = frameCountMax
+
+        while 0 < framesLeft {
+            var frameCount = framesLeft
+            let areas = try! outstream.beginWriting(theNumberOf: &frameCount)
+
+            if frameCount == 0 {
+                break
+            }
+
             let pitch: Float = 440.0
             let radiansPerSecond = pitch * 2.0 * .pi
             for frame in 0..<frameCount {
@@ -28,6 +37,12 @@ func main() throws {
                 }
             }
             secondsOffset = (secondsOffset + secondsPerFrame * Float(frameCount)).truncatingRemainder(dividingBy: 1)
+            try! outstream.endWrite()
+
+            framesLeft -= frameCountMax
+            if framesLeft <= 0 {
+                break
+            }
         }
     }